Move BuildModel.py & fix loading from examples dir
