+++ examples/example_mnist_torch.py
@@ -1,6 +1,5 @@
-
 from models.Models import make_mnist_torch_model
-from example_mnist import *
+from examples.example_mnist import *
 
 get_model = make_mnist_torch_model
